Dernier commit avant commit final

+++ CPOA TD2/src/projet/controleur/PageCategorieController.java
@@ -2,6 +2,8 @@ package projet.controleur;
 
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.Optional;
+
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -9,15 +11,19 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
+import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.stage.Stage;
 import projet.dao.DAOFactory;
 import projet.dao.Persistance;
 import projet.metier.Categorie;
 import projet.metier.Client;
+import projet.metier.LigneCommande;
 
 public class PageCategorieController {
 	
@@ -47,22 +53,32 @@ public class PageCategorieController {
 		Stage window = (Stage)((Node)e.getSource()).getScene().getWindow();
 		window.setScene(addCategoriescene);
 		window.centerOnScreen();
-		window.setTitle("Categorie");
+		window.setTitle("Ajout d'une catégorie");
 		window.show();
 	}
 	
 	@FXML public void clickOnDelete(ActionEvent e) throws IOException, SQLException {
 		DAOFactory dao = DAOFactory.getDAOFactory(Persistance.MYSQL);
-        ObservableList<Categorie> Allpeople, Ligneselect;
-        Allpeople = tableViewCategorie.getItems();
-        Ligneselect = tableViewCategorie.getSelectionModel().getSelectedItems();
-        
-        for(Categorie categorie : Ligneselect)
-        {
-            Allpeople.remove(categorie);
-            System.out.println(categorie);
-            dao.getCategorieDAO().delete(categorie);
-        }     
+		ObservableList<Categorie> Allpeople, Ligneselect;
+		Allpeople = tableViewCategorie.getItems();
+		Ligneselect = tableViewCategorie.getSelectionModel().getSelectedItems();
+
+		Alert alert = new Alert(AlertType.CONFIRMATION);
+		alert.setTitle("Suppresion d'une catégorie");
+		alert.setHeaderText("Êtes vous sur de vouloir supprimer cette catégorie ?");
+
+		Optional<ButtonType> option = alert.showAndWait();
+		if (option.get() == null) {
+		} else if (option.get() == ButtonType.OK) {
+			for(Categorie categorie : Ligneselect)
+			{
+				Allpeople.remove(categorie);
+				System.out.println(categorie);
+				dao.getCategorieDAO().delete(categorie);
+			}
+		}
+		else if (option.get() == ButtonType.CANCEL) {
+		}
 	}
 	
 	@FXML public void clickOnEdit(ActionEvent e) throws IOException {
@@ -72,7 +88,7 @@ public class PageCategorieController {
 		Stage window = (Stage)((Node)e.getSource()).getScene().getWindow();
 		window.setScene(editCategoriescene);
 		window.centerOnScreen();
-		window.setTitle("Categorie");
+		window.setTitle("Modification d'une catégorie");
 		window.show();
 	}
 	
@@ -82,7 +98,7 @@ public class PageCategorieController {
 		Stage window = (Stage)((Node)e.getSource()).getScene().getWindow();
 		window.setScene(returnCategoriescene);
 		window.centerOnScreen();
-		window.setTitle("Categorie");
+		window.setTitle("Application de gestion commercial");
 		window.show();
 	}
